+++ Sources/Frontend/Commands/FrontendCommand.swift
@@ -1,0 +1,7 @@
+import ArgumentParser
+import Shared
+
+protocol FrontendCommand: ParsableCommand {}
+extension FrontendCommand {
+    static var _errorLabel: String { colorize("error", .boldRed) }
+}