+++ Sources/Shared/PeripheryError.swift
@@ -19,6 +19,7 @@ public enum PeripheryError: Error, LocalizedError, CustomStringConvertible {
     case foundIssues(count: Int)
     case packageError(message: String)
     case swiftVersionParseError(fullVersion: String)
+    case unindexedTargetsError(targets: Set<String>, indexStorePath: String)
 
     public var errorDescription: String? {
         switch self {
@@ -57,6 +58,9 @@ public enum PeripheryError: Error, LocalizedError, CustomStringConvertible {
             return message
         case .swiftVersionParseError(let fullVersion):
             return "Failed to parse Swift version from: \(fullVersion)"
+        case let .unindexedTargetsError(targets, indexStorePath):
+            let joinedTargets = targets.sorted().joined(separator: ", ")
+            return "The index store at '\(indexStorePath)' does not contain data for the following targets: \(joinedTargets). Either the index store is outdated, or you have requested to scan targets that have not been built."
         }
     }