+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ##### Enhancements
 
+- Using an index store that does not contain complete data for the requested targets now results in an error.
 - The '--index-store-path' option now implies '--skip-build'.
 
 ##### Bug Fixes